+++ src/lib/Stores.ts
@@ -110,7 +110,8 @@ export const demo = writable<{ [key: string]: string | undefined }>({
 	bar: undefined,
 	radial: undefined,
 	camera: undefined,
-	history: undefined
+	history: undefined,
+	weather_forecast: undefined
 });
 
 // youtube_addon